requirements.txt
PyMuPDF
pdfplumber
sentence-transformers
scikit-learn
tqdm
llama-cpp-python